Initial version of CPack-based Debian packaging
This feature branch implements CPack-based Debian packaging to
ease distribution of binaries. Dysco builds can be optimized for
the native CPU (default), or for a generic x86-64 processor. This
is influenced by the CMake variable PORTABLE. If generating Debian
packages is enabled (CMake variable BUILD_PACKAGES), and if
building for the native CPU, then the supported AVX-features are
encoded in the name of the package.