Add possibility to set target CPU architecture
This PR adds the possibility to specify for which CPU architecture the compiler should build binaries. It replaces the existing method where one or more AVX extensions could be enabled by setting the associated variables. The rationale is that it makes more sense to optimize for a certain CPU micro-architecture (and its successors) than to only enable specific AVX features.

The `PORTABLE` option is retained and its behavior doesn't change. I.e., setting `PORTABLE=ON` will instruct the compiler to build binaries that can be run on any X86-64 CPU; setting `PORTABLE=OFF` will instruct the compiler to use the native CPU as target. The user can instruct the compiler to target a specific CPU by setting the CMake variable `TARGET_CPU`. Setting `TARGET_CPU` if `PORTABLE=ON` will raise an error. Setting `TARGET_CPU` to a CPU type the compiler doesn't support will also raise an error.